add block notify callback in Service.java

+++ src/main/java/org/fisco/bcos/channel/handler/ConnectionCallback.java
@@ -122,14 +122,14 @@ public class ConnectionCallback implements ChannelConnections.Callback {
                                             .put(
                                                     socketAddress.getAddress().getHostAddress()
                                                             + socketAddress.getPort(),
-                                                    blockNumber.getBlockNumber().intValue());
+                                                    blockNumber.getBlockNumber());
                                 } catch (Exception e) {
                                     throw new MessageDecodingException(response.getContent());
                                 }
                             }
                         });
     }
-
+    
     @Override
     public void onDisconnect(ChannelHandlerContext ctx) {}
